+++ R/hal.R
@@ -130,6 +130,11 @@
 #'   \code{\link[glmnet]{cv.glmnet}}. When \code{TRUE}, \code{"lambda.min"} is
 #'   used; otherwise, \code{"lambda.1se"}. Only used when
 #'   \code{cv_select = TRUE}.
+#' - \code{lambda.min.ratio}: A \code{\link[glmnet]{glmnet}} argument specifying
+#'   the smallest value for \code{lambda}, as a fraction of \code{lambda.max},
+#'   the (data derived) entry value (i.e. the smallest value for which all
+#'   coefficients are zero). We've seen that not setting \code{lambda.min.ratio}
+#'   can lead to no \code{lambda} values that fit the data sufficiently well.
 #' - \code{prediction_bounds}: A vector of size two that provides the lower and
 #'   upper bounds for predictions. When \code{prediction_bounds = "default"},
 #'   the predictions are bounded between \code{min(Y) - sd(Y)} and
@@ -173,8 +178,8 @@ fit_hal <- function(X,
                     X_unpenalized = NULL,
                     max_degree = ifelse(ncol(X) >= 20, 2, 3),
                     smoothness_orders = 1,
-                    num_knots = sapply(seq_len(max_degree),
-                      num_knots_generator,
+                    num_knots = num_knots_generator(
+                      max_degree = max_degree,
                       smoothness_orders = smoothness_orders,
                       base_num_knots_0 = 500,
                       base_num_knots_1 = 200
@@ -189,6 +194,7 @@ fit_hal <- function(X,
                       n_folds = 10,
                       foldid = NULL,
                       use_min = TRUE,
+                      lambda.min.ratio = 1e-4,
                       prediction_bounds = "default"
                     ),
                     formula_control = list(
@@ -206,7 +212,7 @@ fit_hal <- function(X,
   # errors when a supplied control list is missing arguments
   defaults <- list(
     cv_select = TRUE, n_folds = 10, foldid = NULL, use_min = TRUE,
-    prediction_bounds = "default"
+    lambda.min.ratio = 1e-4, prediction_bounds = "default"
   )
   if (any(!names(defaults) %in% names(fit_control))) {
     fit_control <- c(
@@ -472,12 +478,16 @@ fit_hal <- function(X,
 #'  the basis function.
 #'
 #' @keywords internal
-num_knots_generator <- function(d, smoothness_orders, base_num_knots_0 = 500,
+num_knots_generator <- function(max_degree, smoothness_orders, base_num_knots_0 = 500,
                                 base_num_knots_1 = 200) {
   if (all(smoothness_orders > 0)) {
-    return(round(base_num_knots_1 / 2^(d - 1)))
+    return(sapply(seq_len(max_degree), function(d) {
+      round(base_num_knots_1 / 2^(d - 1))
+    }))
   }
   else {
-    return(round(base_num_knots_0 / 2^(d - 1)))
+    return(sapply(seq_len(max_degree), function(d) {
+      round(base_num_knots_0 / 2^(d - 1))
+    }))
   }
 }