+++ tests/testthat/test-general_families.R
@@ -20,7 +20,7 @@ test_y_prob <- plogis(3 * sin(test_x[, 1]) + sin(test_x[, 2]))
 test_y <- rbinom(n = test_n, size = 1, prob = y_prob)
 
 # ml implementation
-ml_hal_fit <- fit_hal(X = x, Y = y, family = "binomial", yolo = FALSE)
+ml_hal_fit <- suppressWarnings(fit_hal(X = x, Y = y, family = "binomial"))
 ml_hal_fit$times
 x_basis <- make_design_matrix(x, ml_hal_fit$basis_list)
 
@@ -29,7 +29,7 @@ preds <- predict(ml_hal_fit, new_data = x)
 ml_hal_mse1 <- mse(preds, y_prob)
 set.seed(45791)
 
-ml_hal_fit <- fit_hal(X = x, Y = y, family = binomial(), yolo = FALSE)
+ml_hal_fit <- suppressWarnings(fit_hal(X = x, Y = y, family = binomial()))
 ml_hal_fit$times
 x_basis <- make_design_matrix(x, ml_hal_fit$basis_list)
 
@@ -44,7 +44,7 @@ test_that("MSE for logistic regression close to logistic family object pred", {
 
 
 # ml implementation
-ml_hal_fit <- fit_hal(X = x, Y = y, family = "poisson", yolo = FALSE)
+ml_hal_fit <- suppressWarnings(fit_hal(X = x, Y = y, family = "poisson"))
 ml_hal_fit$times
 x_basis <- make_design_matrix(x, ml_hal_fit$basis_list)
 
@@ -53,7 +53,7 @@ preds <- predict(ml_hal_fit, new_data = x)
 ml_hal_mse1 <- mse(preds, y_prob)
 set.seed(45791)
 
-ml_hal_fit <- fit_hal(X = x, Y = y, family = poisson(), yolo = FALSE)
+ml_hal_fit <- suppressWarnings(fit_hal(X = x, Y = y, family = poisson()))
 ml_hal_fit$times
 x_basis <- make_design_matrix(x, ml_hal_fit$basis_list)